Refactor job_output_to_atomic_result function to be client method to dynamically set AtomicResult.provenance for depending on client used.
